23.03

+++ app/src/main/java/com/example/snav1/Product.kt
@@ -6,7 +6,7 @@ import java.io.Serializable
 
 class Product(var id:Int=0,
               var name:String?=null,
-              var price:Double=0.0,
+              var price:Float=0.0f,
               var type:String?=null,
               var img:Int,
               var description:String?= "Ürüne detaylı açıklaması", ) :Serializable{
