+++ app/src/main/java/com/example/snav1/card/cardTotal.kt
@@ -1,3 +1,3 @@
 package com.example.snav1.card
 
-object cardTotal: ArrayList<Double>() { }
+object cardTotal: ArrayList<Float>() { }
